+++ src/test/samples/csharp.cs
@@ -1,0 +1,28 @@
+namespace BetterComments
+{
+    /// <summary>
+    /// The CSharpSample class.
+    /// </summary>
+    public class CSharpSample
+    {
+        public int Count = 0;
+
+        public CSharpSample()
+        {
+        }
+
+        public IncreaseCount()
+        {
+            // TODO implement a method that adds to the count
+            // ! single line comments are highlighted also
+
+            /**
+             * ! alerts can be in multilines
+             */
+
+             /*
+              ! you don't need a preceeding * all the time, only when block comments begin with /**
+              */
+        }
+    }
+}